Added nearby search to be based off of user's location (to-do, make nearby search based off of delivery address).

+++ app/src/main/java/com/codepath/tsazo/consumer/fragments/DriverOrderFragment.java
@@ -216,7 +216,7 @@ public class DriverOrderFragment extends Fragment {
             public void onClick(View view) {
                 // TODO: DO a check to see if Driver is at/near delivery address before completing order
                 currentUser.put(KEY_HAS_ORDER, false);
-                Number totalEarnings = (float) currentUser.getNumber(KEY_EARNINGS) + (float) order.getPrice();
+                Number totalEarnings = (Integer) currentUser.getNumber(KEY_EARNINGS) + (Integer) order.getPrice();
                 currentUser.put(KEY_EARNINGS, totalEarnings);
                 order.setIsDone(true);
